Updates to the interface for better call to action
Call to action added along with the hiding and showing of the meta data upon playing and pausing.

Also corrected a bug where music was auto starting by default and causing dom exceptions. It now requires an action by the user to start.

Changing the bitrate before a stream does NOT instigate an auto play however changing it after playback has happened results in it remaining its last state continuing.